requirements.txt
httpx
openai>=1.40.0
PyYAML
python-dotenv
llama-cpp-python; platform_machine != "arm64" or sys_platform != "darwin"
fastapi>=0.110
uvicorn[standard]>=0.24
